requirements.txt
flask
flask_cors
paho-mqtt
requests